backend/requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
setuptools==69.5.1
werkzeug==3.0.3
wheel==0.43.0
pymysql[rsa]
python-dotenv
requests
bs4
uuid
flask-cors
pyjwt
bcrypt
pillow
segno
python-dateutil